requirements.txt
tensorflow-gpu
keras-rl
gym
matplotlib